% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Files2Vec}
\alias{Files2Vec}
\title{Function for reading text files into R}
\usage{
Files2Vec(directory)
}
\arguments{
\item{directory}{a path to directory containing the files you want to read into R.}
}
\description{
This function reads in all files ending in .txt into R.
The result is a character vector where each entry is a .txt file.
The names of the resulting vector are derived from the names of the files.
}
\examples{
my_text_vector <- Files2Vec("//div-stpi/public/___STPI_Tools/DocSummarizer/data/")
}

