% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ImportDocs}
\alias{ImportDocs}
\title{Import Documents for IDA's DTM-making Shiny app.}
\usage{
ImportDocs(user.input)
}
\arguments{
\item{user.input}{= a list with elements directing the
  method for creating a DTM from files located in a directory}
}
\description{
This is the main document importing function. It is formatted
to work in the confines of IDA's DTM-making Shiny app.
Those making an R script should *not*
use this function. Instead, use other functions, such as
Vec2Dtm() directly.
}
\details{
output: a list with two objects
      dtm = the document term matrix
      tf.mat = a dataframe of corpus-wide statistics
}

